+++ include/mull/Mutators/ReplaceAssignmentMutator.h
@@ -18,6 +18,9 @@ public:
   std::string getUniqueIdentifier() override { return ID; }
   std::string getUniqueIdentifier() const override { return ID; }
   std::string getDescription() const override { return description; }
+  MutatorKind mutatorKind() override {
+    return MutatorKind::ReplaceAssignmentMutator;
+  }
 
   void applyMutation(llvm::Function *function,
                      const MutationPointAddress &address) override;